Fix build: wrap not-found page in Suspense boundary
Next.js 16 Turbopack requires client components using useRouter to be
wrapped in Suspense for static prerendering. Split not-found.tsx into
a server component wrapper + client content component.
